Added a MCTS-based AI with heuristics.
An AI with MCTS algorithm as its core logic, with additional heuristics more sophisticated compared to the AI in main_grok. However, this implementation of MCTS includes a Zobrist hashing transposition table to reduce computational resource, thus time per move. Does not help that much though, as it is still the weakest of the bunch.